+++ source_files/edge/s_prime.cc
@@ -68,7 +68,7 @@ bool S_StartupPrime(void)
 
 	// Check for presence of previous CVAR value's file
 	bool cvar_good = false;
-	for (int i=0; i < available_soundfonts.size(); i++)
+	for (size_t i=0; i < available_soundfonts.size(); i++)
 	{
 		if(epi::case_cmp(s_soundfont.s, available_soundfonts.at(i).generic_u8string()) == 0)
 		{
